Support partial metadata based aggregation when some projection aggregations are not metadata compatible
