+++ pinot-core/src/main/java/org/apache/pinot/core/operator/query/NonScanBasedAggregationOperator.java
@@ -18,43 +18,16 @@
  */
 package org.apache.pinot.core.operator.query;
 
-import com.clearspring.analytics.stream.cardinality.HyperLogLog;
-import com.clearspring.analytics.stream.cardinality.HyperLogLogPlus;
-import com.dynatrace.hash4j.distinctcount.UltraLogLog;
-import com.google.common.base.Preconditions;
-import it.unimi.dsi.fastutil.doubles.DoubleOpenHashSet;
-import it.unimi.dsi.fastutil.floats.FloatOpenHashSet;
-import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
-import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
-import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-import java.util.Set;
-import org.apache.pinot.core.common.ObjectSerDeUtils;
 import org.apache.pinot.core.common.Operator;
 import org.apache.pinot.core.operator.BaseOperator;
 import org.apache.pinot.core.operator.ExecutionStatistics;
 import org.apache.pinot.core.operator.blocks.results.AggregationResultsBlock;
 import org.apache.pinot.core.query.aggregation.function.AggregationFunction;
-import org.apache.pinot.core.query.aggregation.function.DistinctCountHLLAggregationFunction;
-import org.apache.pinot.core.query.aggregation.function.DistinctCountHLLPlusAggregationFunction;
-import org.apache.pinot.core.query.aggregation.function.DistinctCountOffHeapAggregationFunction;
-import org.apache.pinot.core.query.aggregation.function.DistinctCountRawHLLAggregationFunction;
-import org.apache.pinot.core.query.aggregation.function.DistinctCountRawHLLPlusAggregationFunction;
-import org.apache.pinot.core.query.aggregation.function.DistinctCountSmartHLLAggregationFunction;
-import org.apache.pinot.core.query.aggregation.function.DistinctCountSmartHLLPlusAggregationFunction;
-import org.apache.pinot.core.query.aggregation.function.DistinctCountSmartULLAggregationFunction;
-import org.apache.pinot.core.query.aggregation.function.DistinctCountULLAggregationFunction;
+import org.apache.pinot.core.query.aggregation.function.AggregationFunctionUtils;
 import org.apache.pinot.core.query.request.context.QueryContext;
-import org.apache.pinot.segment.local.customobject.MinMaxRangePair;
-import org.apache.pinot.segment.local.utils.UltraLogLogUtils;
 import org.apache.pinot.segment.spi.datasource.DataSource;
-import org.apache.pinot.segment.spi.index.reader.Dictionary;
-import org.apache.pinot.spi.data.FieldSpec.DataType;
-import org.apache.pinot.spi.query.QueryThreadContext;
-import org.apache.pinot.spi.utils.ByteArray;
 
 
 /// Aggregation operator that utilizes dictionary or column metadata for serving aggregation queries to avoid scanning.
@@ -72,7 +45,7 @@ import org.apache.pinot.spi.utils.ByteArray;
 /// is no dictionary.
 @SuppressWarnings("rawtypes")
 public class NonScanBasedAggregationOperator extends BaseOperator<AggregationResultsBlock> {
-  private static final String EXPLAIN_NAME = "AGGREGATE_NO_SCAN";
+  public static final String EXPLAIN_NAME = "AGGREGATE_NO_SCAN";
 
   private final QueryContext _queryContext;
   private final AggregationFunction[] _aggregationFunctions;
@@ -93,332 +66,13 @@ public class NonScanBasedAggregationOperator extends BaseOperator<AggregationRes
       AggregationFunction aggregationFunction = _aggregationFunctions[i];
       // note that dataSource will be null for COUNT, sp do not interact with it until it's known this isn't a COUNT
       DataSource dataSource = _dataSources[i];
-      Object result;
-      switch (aggregationFunction.getType()) {
-        case COUNT:
-          result = (long) _numTotalDocs;
-          break;
-        case MIN:
-        case MINMV:
-          result = getMinValueNumeric(dataSource);
-          break;
-        case MINLONG:
-          result = getMinValueLong(dataSource);
-          break;
-        case MINSTRING:
-          assert dataSource.getDictionary() != null;
-          result = dataSource.getDictionary().getMinVal();
-          break;
-        case MAX:
-        case MAXMV:
-          result = getMaxValueNumeric(dataSource);
-          break;
-        case MAXLONG:
-          result = getMaxValueLong(dataSource);
-          break;
-        case MAXSTRING:
-          assert dataSource.getDictionary() != null;
-          result = dataSource.getDictionary().getMaxVal();
-          break;
-        case MINMAXRANGE:
-        case MINMAXRANGEMV:
-          result = new MinMaxRangePair(getMinValueNumeric(dataSource), getMaxValueNumeric(dataSource));
-          break;
-        case DISTINCTCOUNT:
-        case DISTINCTSUM:
-        case DISTINCTAVG:
-        case DISTINCTCOUNTMV:
-        case DISTINCTSUMMV:
-        case DISTINCTAVGMV:
-          result = getDistinctValueSet(Objects.requireNonNull(dataSource.getDictionary()));
-          break;
-        case DISTINCTCOUNTOFFHEAP:
-          result = ((DistinctCountOffHeapAggregationFunction) aggregationFunction).extractAggregationResult(
-              Objects.requireNonNull(dataSource.getDictionary()));
-          break;
-        case DISTINCTCOUNTHLL:
-        case DISTINCTCOUNTHLLMV:
-          result = getDistinctCountHLLResult(Objects.requireNonNull(dataSource.getDictionary()),
-              (DistinctCountHLLAggregationFunction) aggregationFunction);
-          break;
-        case DISTINCTCOUNTRAWHLL:
-        case DISTINCTCOUNTRAWHLLMV:
-          result = getDistinctCountHLLResult(Objects.requireNonNull(dataSource.getDictionary()),
-              ((DistinctCountRawHLLAggregationFunction) aggregationFunction).getDistinctCountHLLAggregationFunction());
-          break;
-        case DISTINCTCOUNTHLLPLUS:
-        case DISTINCTCOUNTHLLPLUSMV:
-          result = getDistinctCountHLLPlusResult(Objects.requireNonNull(dataSource.getDictionary()),
-              (DistinctCountHLLPlusAggregationFunction) aggregationFunction);
-          break;
-        case DISTINCTCOUNTRAWHLLPLUS:
-        case DISTINCTCOUNTRAWHLLPLUSMV:
-          result = getDistinctCountHLLPlusResult(Objects.requireNonNull(dataSource.getDictionary()),
-              ((DistinctCountRawHLLPlusAggregationFunction) aggregationFunction)
-                  .getDistinctCountHLLPlusAggregationFunction());
-          break;
-        case SEGMENTPARTITIONEDDISTINCTCOUNT:
-          result = (long) Objects.requireNonNull(dataSource.getDictionary()).length();
-          break;
-        case DISTINCTCOUNTSMARTHLL:
-          result = getDistinctCountSmartHLLResult(Objects.requireNonNull(dataSource.getDictionary()),
-              (DistinctCountSmartHLLAggregationFunction) aggregationFunction);
-          break;
-        case DISTINCTCOUNTSMARTHLLPLUS:
-          result = getDistinctCountSmartHLLPlusResult(Objects.requireNonNull(dataSource.getDictionary()),
-              (DistinctCountSmartHLLPlusAggregationFunction) aggregationFunction);
-          break;
-        case DISTINCTCOUNTULL:
-          result = getDistinctCountULLResult(Objects.requireNonNull(dataSource.getDictionary()),
-              (DistinctCountULLAggregationFunction) aggregationFunction);
-          break;
-        case DISTINCTCOUNTSMARTULL:
-          result = getDistinctCountSmartULLResult(Objects.requireNonNull(dataSource.getDictionary()),
-              (DistinctCountSmartULLAggregationFunction) aggregationFunction);
-          break;
-        case DISTINCTCOUNTRAWULL:
-          result = getDistinctCountULLResult(Objects.requireNonNull(dataSource.getDictionary()),
-              (DistinctCountULLAggregationFunction) aggregationFunction);
-          break;
-        default:
-          throw new IllegalStateException(
-              "Non-scan based aggregation operator does not support function type: " + aggregationFunction.getType());
-      }
+      Object result = AggregationFunctionUtils.getAggregationResult(aggregationFunction, dataSource,
+          _numTotalDocs, EXPLAIN_NAME);
       aggregationResults.add(result);
     }
 
     // Build intermediate result block based on aggregation result from the executor.
     return new AggregationResultsBlock(_aggregationFunctions, aggregationResults, _queryContext);
-  }
-
-  private static Double getMinValueNumeric(DataSource dataSource) {
-    Dictionary dictionary = dataSource.getDictionary();
-    if (dictionary != null) {
-      return toDouble(dictionary.getMinVal());
-    }
-    return toDouble(dataSource.getDataSourceMetadata().getMinValue());
-  }
-
-  private static Long getMinValueLong(DataSource dataSource) {
-    DataType dataType = dataSource.getDataSourceMetadata().getDataType().getStoredType();
-    Preconditions.checkArgument(
-        dataType == DataType.LONG || dataType == DataType.INT,
-        "MINLONG aggregation function can only be applied to columns of integer types");
-    Dictionary dictionary = dataSource.getDictionary();
-    if (dictionary != null) {
-      return ((Number) dictionary.getMinVal()).longValue();
-    }
-    return ((Number) dataSource.getDataSourceMetadata().getMinValue()).longValue();
-  }
-
-  private static Double getMaxValueNumeric(DataSource dataSource) {
-    Dictionary dictionary = dataSource.getDictionary();
-    if (dictionary != null) {
-      return toDouble(dictionary.getMaxVal());
-    }
-    return toDouble(dataSource.getDataSourceMetadata().getMaxValue());
-  }
-
-  private static Long getMaxValueLong(DataSource dataSource) {
-    DataType dataType = dataSource.getDataSourceMetadata().getDataType().getStoredType();
-    Preconditions.checkArgument(
-        dataType == DataType.LONG || dataType == DataType.INT,
-        "MAXLONG aggregation function can only be applied to columns of integer types");
-    Dictionary dictionary = dataSource.getDictionary();
-    if (dictionary != null) {
-      return ((Number) dictionary.getMaxVal()).longValue();
-    }
-    return ((Number) dataSource.getDataSourceMetadata().getMaxValue()).longValue();
-  }
-
-  private static Double toDouble(Comparable<?> value) {
-    if (value instanceof Double) {
-      return (Double) value;
-    } else if (value instanceof Number) {
-      return ((Number) value).doubleValue();
-    } else {
-      return Double.parseDouble(value.toString());
-    }
-  }
-
-  private static Set getDistinctValueSet(Dictionary dictionary) {
-    int dictionarySize = dictionary.length();
-    switch (dictionary.getValueType()) {
-      case INT:
-        IntOpenHashSet intSet = new IntOpenHashSet(dictionarySize);
-        for (int dictId = 0; dictId < dictionarySize; dictId++) {
-          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, EXPLAIN_NAME);
-          intSet.add(dictionary.getIntValue(dictId));
-        }
-        return intSet;
-      case LONG:
-        LongOpenHashSet longSet = new LongOpenHashSet(dictionarySize);
-        for (int dictId = 0; dictId < dictionarySize; dictId++) {
-          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, EXPLAIN_NAME);
-          longSet.add(dictionary.getLongValue(dictId));
-        }
-        return longSet;
-      case FLOAT:
-        FloatOpenHashSet floatSet = new FloatOpenHashSet(dictionarySize);
-        for (int dictId = 0; dictId < dictionarySize; dictId++) {
-          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, EXPLAIN_NAME);
-          floatSet.add(dictionary.getFloatValue(dictId));
-        }
-        return floatSet;
-      case DOUBLE:
-        DoubleOpenHashSet doubleSet = new DoubleOpenHashSet(dictionarySize);
-        for (int dictId = 0; dictId < dictionarySize; dictId++) {
-          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, EXPLAIN_NAME);
-          doubleSet.add(dictionary.getDoubleValue(dictId));
-        }
-        return doubleSet;
-      case BIG_DECIMAL:
-        ObjectOpenHashSet<BigDecimal> bigDecimalSet = new ObjectOpenHashSet<>(dictionarySize);
-        for (int dictId = 0; dictId < dictionarySize; dictId++) {
-          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, EXPLAIN_NAME);
-          bigDecimalSet.add(dictionary.getBigDecimalValue(dictId));
-        }
-        return bigDecimalSet;
-      case STRING:
-        ObjectOpenHashSet<String> stringSet = new ObjectOpenHashSet<>(dictionarySize);
-        for (int dictId = 0; dictId < dictionarySize; dictId++) {
-          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, EXPLAIN_NAME);
-          stringSet.add(dictionary.getStringValue(dictId));
-        }
-        return stringSet;
-      case BYTES:
-        ObjectOpenHashSet<ByteArray> bytesSet = new ObjectOpenHashSet<>(dictionarySize);
-        for (int dictId = 0; dictId < dictionarySize; dictId++) {
-          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, EXPLAIN_NAME);
-          bytesSet.add(new ByteArray(dictionary.getBytesValue(dictId)));
-        }
-        return bytesSet;
-      default:
-        throw new IllegalStateException();
-    }
-  }
-
-  private static HyperLogLog getDistinctValueHLL(Dictionary dictionary, int log2m) {
-    HyperLogLog hll = new HyperLogLog(log2m);
-    int length = dictionary.length();
-    for (int i = 0; i < length; i++) {
-      QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, EXPLAIN_NAME);
-      hll.offer(dictionary.get(i));
-    }
-    return hll;
-  }
-
-  private static UltraLogLog getDistinctValueULL(Dictionary dictionary, int p) {
-    UltraLogLog ull = UltraLogLog.create(p);
-    int length = dictionary.length();
-    for (int i = 0; i < length; i++) {
-      QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, EXPLAIN_NAME);
-      Object value = dictionary.get(i);
-      UltraLogLogUtils.hashObject(value).ifPresent(ull::add);
-    }
-    return ull;
-  }
-
-  private static HyperLogLogPlus getDistinctValueHLLPlus(Dictionary dictionary, int p, int sp) {
-    HyperLogLogPlus hllPlus = new HyperLogLogPlus(p, sp);
-    int length = dictionary.length();
-    for (int i = 0; i < length; i++) {
-      QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, EXPLAIN_NAME);
-      hllPlus.offer(dictionary.get(i));
-    }
-    return hllPlus;
-  }
-
-  private static HyperLogLog getDistinctCountHLLResult(Dictionary dictionary,
-      DistinctCountHLLAggregationFunction function) {
-    if (dictionary.getValueType() == DataType.BYTES) {
-      // Treat BYTES value as serialized HyperLogLog
-      try {
-        QueryThreadContext.checkTerminationAndSampleUsage(EXPLAIN_NAME);
-        HyperLogLog hll = ObjectSerDeUtils.HYPER_LOG_LOG_SER_DE.deserialize(dictionary.getBytesValue(0));
-        int length = dictionary.length();
-        for (int i = 1; i < length; i++) {
-          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, EXPLAIN_NAME);
-          hll.addAll(ObjectSerDeUtils.HYPER_LOG_LOG_SER_DE.deserialize(dictionary.getBytesValue(i)));
-        }
-        return hll;
-      } catch (Exception e) {
-        throw new RuntimeException("Caught exception while merging HyperLogLogs", e);
-      }
-    } else {
-      return getDistinctValueHLL(dictionary, function.getLog2m());
-    }
-  }
-
-  private static HyperLogLogPlus getDistinctCountHLLPlusResult(Dictionary dictionary,
-      DistinctCountHLLPlusAggregationFunction function) {
-    if (dictionary.getValueType() == DataType.BYTES) {
-      // Treat BYTES value as serialized HyperLogLogPlus
-      try {
-        QueryThreadContext.checkTerminationAndSampleUsage(EXPLAIN_NAME);
-        HyperLogLogPlus hllplus = ObjectSerDeUtils.HYPER_LOG_LOG_PLUS_SER_DE.deserialize(dictionary.getBytesValue(0));
-        int length = dictionary.length();
-        for (int i = 1; i < length; i++) {
-          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, EXPLAIN_NAME);
-          hllplus.addAll(ObjectSerDeUtils.HYPER_LOG_LOG_PLUS_SER_DE.deserialize(dictionary.getBytesValue(i)));
-        }
-        return hllplus;
-      } catch (Exception e) {
-        throw new RuntimeException("Caught exception while merging HyperLogLogPluses", e);
-      }
-    } else {
-      return getDistinctValueHLLPlus(dictionary, function.getP(), function.getSp());
-    }
-  }
-
-  private static Object getDistinctCountSmartHLLResult(Dictionary dictionary,
-      DistinctCountSmartHLLAggregationFunction function) {
-    if (dictionary.length() > function.getThreshold()) {
-      // Store values into a HLL when the dictionary size exceeds the conversion threshold
-      return getDistinctValueHLL(dictionary, function.getLog2m());
-    } else {
-      return getDistinctValueSet(dictionary);
-    }
-  }
-
-  private static Object getDistinctCountSmartHLLPlusResult(Dictionary dictionary,
-      DistinctCountSmartHLLPlusAggregationFunction function) {
-    if (dictionary.length() > function.getThreshold()) {
-      // Store values into a HLLPlus when the dictionary size exceeds the conversion threshold
-      return getDistinctValueHLLPlus(dictionary, function.getP(), function.getSp());
-    } else {
-      return getDistinctValueSet(dictionary);
-    }
-  }
-
-  private static UltraLogLog getDistinctCountULLResult(Dictionary dictionary,
-      DistinctCountULLAggregationFunction function) {
-    if (dictionary.getValueType() == DataType.BYTES) {
-      // Treat BYTES value as serialized UltraLogLog and merge
-      try {
-        QueryThreadContext.checkTerminationAndSampleUsage(EXPLAIN_NAME);
-        UltraLogLog ull = ObjectSerDeUtils.ULTRA_LOG_LOG_OBJECT_SER_DE.deserialize(dictionary.getBytesValue(0));
-        int length = dictionary.length();
-        for (int i = 1; i < length; i++) {
-          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, EXPLAIN_NAME);
-          ull.add(ObjectSerDeUtils.ULTRA_LOG_LOG_OBJECT_SER_DE.deserialize(dictionary.getBytesValue(i)));
-        }
-        return ull;
-      } catch (Exception e) {
-        throw new RuntimeException("Caught exception while merging UltraLogLogs", e);
-      }
-    } else {
-      return getDistinctValueULL(dictionary, function.getP());
-    }
-  }
-
-  private static Object getDistinctCountSmartULLResult(Dictionary dictionary,
-      DistinctCountSmartULLAggregationFunction function) {
-    if (dictionary.length() > function.getThreshold()) {
-      return getDistinctValueULL(dictionary, function.getP());
-    } else {
-      return getDistinctValueSet(dictionary);
-    }
   }
 
   @Override
