+++ pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/function/AggregationFunctionUtils.java
@@ -18,11 +18,20 @@
  */
 package org.apache.pinot.core.query.aggregation.function;
 
+import com.clearspring.analytics.stream.cardinality.HyperLogLog;
+import com.clearspring.analytics.stream.cardinality.HyperLogLogPlus;
+import com.dynatrace.hash4j.distinctcount.UltraLogLog;
+import com.google.common.base.Preconditions;
 import it.unimi.dsi.fastutil.doubles.DoubleArrayList;
+import it.unimi.dsi.fastutil.doubles.DoubleOpenHashSet;
 import it.unimi.dsi.fastutil.floats.FloatArrayList;
+import it.unimi.dsi.fastutil.floats.FloatOpenHashSet;
 import it.unimi.dsi.fastutil.ints.IntArrayList;
+import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
 import it.unimi.dsi.fastutil.longs.LongArrayList;
+import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
+import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.sql.Timestamp;
@@ -31,6 +40,7 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
 import javax.annotation.Nullable;
 import org.apache.commons.lang3.tuple.Pair;
@@ -42,6 +52,7 @@ import org.apache.pinot.common.request.context.predicate.Predicate;
 import org.apache.pinot.common.utils.DataSchema.ColumnDataType;
 import org.apache.pinot.common.utils.config.QueryOptionsUtils;
 import org.apache.pinot.core.common.BlockValSet;
+import org.apache.pinot.core.common.ObjectSerDeUtils;
 import org.apache.pinot.core.common.datatable.DataTableBuilder;
 import org.apache.pinot.core.operator.BaseProjectOperator;
 import org.apache.pinot.core.operator.blocks.ValueBlock;
@@ -53,15 +64,22 @@ import org.apache.pinot.core.plan.FilterPlanNode;
 import org.apache.pinot.core.plan.ProjectPlanNode;
 import org.apache.pinot.core.query.request.context.QueryContext;
 import org.apache.pinot.core.startree.StarTreeUtils;
+import org.apache.pinot.segment.local.customobject.MinMaxRangePair;
+import org.apache.pinot.segment.local.utils.UltraLogLogUtils;
 import org.apache.pinot.segment.spi.AggregationFunctionType;
 import org.apache.pinot.segment.spi.SegmentContext;
+import org.apache.pinot.segment.spi.datasource.DataSource;
+import org.apache.pinot.segment.spi.index.reader.Dictionary;
 import org.apache.pinot.segment.spi.index.startree.AggregationFunctionColumnPair;
+import org.apache.pinot.spi.data.FieldSpec;
+import org.apache.pinot.spi.query.QueryThreadContext;
 import org.apache.pinot.spi.utils.ByteArray;
 
 
 /// The `AggregationFunctionUtils` class provides utility methods for aggregation function.
 @SuppressWarnings({"rawtypes", "unchecked"})
 public class AggregationFunctionUtils {
+
   private AggregationFunctionUtils() {
   }
 
@@ -476,5 +494,341 @@ public class AggregationFunctionUtils {
       columnName += " FILTER(WHERE " + filter + ")";
     }
     return columnName;
+  }
+
+  /**
+   * Gets the aggregation result without scanning the segment.
+   * This is used for non-scan based aggregation operator.
+   * @param aggregationFunction
+   * @param dataSource
+   * @param numTotalDocs
+   * @return
+   */
+  public static Object getAggregationResult(AggregationFunction aggregationFunction, DataSource dataSource,
+      int numTotalDocs, String explainName) {
+    Object result;
+    switch (aggregationFunction.getType()) {
+      case COUNT:
+        result = (long) numTotalDocs;
+        break;
+      case MIN:
+      case MINMV:
+        result = getMinValueNumeric(dataSource);
+        break;
+      case MINLONG:
+        result = getMinValueLong(dataSource);
+        break;
+      case MINSTRING:
+        assert dataSource.getDictionary() != null;
+        result = dataSource.getDictionary().getMinVal();
+        break;
+      case MAX:
+      case MAXMV:
+        result = getMaxValueNumeric(dataSource);
+        break;
+      case MAXLONG:
+        result = getMaxValueLong(dataSource);
+        break;
+      case MAXSTRING:
+        assert dataSource.getDictionary() != null;
+        result = dataSource.getDictionary().getMaxVal();
+        break;
+      case MINMAXRANGE:
+      case MINMAXRANGEMV:
+        result = new MinMaxRangePair(getMinValueNumeric(dataSource), getMaxValueNumeric(dataSource));
+        break;
+      case DISTINCTCOUNT:
+      case DISTINCTSUM:
+      case DISTINCTAVG:
+      case DISTINCTCOUNTMV:
+      case DISTINCTSUMMV:
+      case DISTINCTAVGMV:
+        result = getDistinctValueSet(Objects.requireNonNull(dataSource.getDictionary()), explainName);
+        break;
+      case DISTINCTCOUNTOFFHEAP:
+        result = ((DistinctCountOffHeapAggregationFunction) aggregationFunction).extractAggregationResult(
+            Objects.requireNonNull(dataSource.getDictionary()));
+        break;
+      case DISTINCTCOUNTHLL:
+      case DISTINCTCOUNTHLLMV:
+        result = getDistinctCountHLLResult(Objects.requireNonNull(dataSource.getDictionary()),
+            (DistinctCountHLLAggregationFunction) aggregationFunction, explainName);
+        break;
+      case DISTINCTCOUNTRAWHLL:
+      case DISTINCTCOUNTRAWHLLMV:
+        result = getDistinctCountHLLResult(Objects.requireNonNull(dataSource.getDictionary()),
+            ((DistinctCountRawHLLAggregationFunction) aggregationFunction).getDistinctCountHLLAggregationFunction(),
+            explainName);
+        break;
+      case DISTINCTCOUNTHLLPLUS:
+      case DISTINCTCOUNTHLLPLUSMV:
+        result = getDistinctCountHLLPlusResult(Objects.requireNonNull(dataSource.getDictionary()),
+            (DistinctCountHLLPlusAggregationFunction) aggregationFunction, explainName);
+        break;
+      case DISTINCTCOUNTRAWHLLPLUS:
+      case DISTINCTCOUNTRAWHLLPLUSMV:
+        result = getDistinctCountHLLPlusResult(Objects.requireNonNull(dataSource.getDictionary()),
+            ((DistinctCountRawHLLPlusAggregationFunction) aggregationFunction)
+                .getDistinctCountHLLPlusAggregationFunction(), explainName);
+        break;
+      case SEGMENTPARTITIONEDDISTINCTCOUNT:
+        result = (long) Objects.requireNonNull(dataSource.getDictionary()).length();
+        break;
+      case DISTINCTCOUNTSMARTHLL:
+        result = getDistinctCountSmartHLLResult(Objects.requireNonNull(dataSource.getDictionary()),
+            (DistinctCountSmartHLLAggregationFunction) aggregationFunction, explainName);
+        break;
+      case DISTINCTCOUNTSMARTHLLPLUS:
+        result = getDistinctCountSmartHLLPlusResult(Objects.requireNonNull(dataSource.getDictionary()),
+            (DistinctCountSmartHLLPlusAggregationFunction) aggregationFunction, explainName);
+        break;
+      case DISTINCTCOUNTULL:
+        result = getDistinctCountULLResult(Objects.requireNonNull(dataSource.getDictionary()),
+            (DistinctCountULLAggregationFunction) aggregationFunction, explainName);
+        break;
+      case DISTINCTCOUNTSMARTULL:
+        result = getDistinctCountSmartULLResult(Objects.requireNonNull(dataSource.getDictionary()),
+            (DistinctCountSmartULLAggregationFunction) aggregationFunction, explainName);
+        break;
+      case DISTINCTCOUNTRAWULL:
+        result = getDistinctCountULLResult(Objects.requireNonNull(dataSource.getDictionary()),
+            (DistinctCountULLAggregationFunction) aggregationFunction, explainName);
+        break;
+      default:
+        throw new IllegalStateException(
+            "Non-scan based aggregation operator does not support function type: " + aggregationFunction.getType());
+    }
+
+    return result;
+  }
+
+  private static Double getMinValueNumeric(DataSource dataSource) {
+    Dictionary dictionary = dataSource.getDictionary();
+    if (dictionary != null) {
+      return toDouble(dictionary.getMinVal());
+    }
+    return toDouble(dataSource.getDataSourceMetadata().getMinValue());
+  }
+
+  private static Long getMinValueLong(DataSource dataSource) {
+    FieldSpec.DataType dataType = dataSource.getDataSourceMetadata().getDataType().getStoredType();
+    Preconditions.checkArgument(
+        dataType == FieldSpec.DataType.LONG || dataType == FieldSpec.DataType.INT,
+        "MINLONG aggregation function can only be applied to columns of integer types");
+    Dictionary dictionary = dataSource.getDictionary();
+    if (dictionary != null) {
+      return ((Number) dictionary.getMinVal()).longValue();
+    }
+    return ((Number) dataSource.getDataSourceMetadata().getMinValue()).longValue();
+  }
+
+  private static Double getMaxValueNumeric(DataSource dataSource) {
+    Dictionary dictionary = dataSource.getDictionary();
+    if (dictionary != null) {
+      return toDouble(dictionary.getMaxVal());
+    }
+    return toDouble(dataSource.getDataSourceMetadata().getMaxValue());
+  }
+
+  private static Long getMaxValueLong(DataSource dataSource) {
+    FieldSpec.DataType dataType = dataSource.getDataSourceMetadata().getDataType().getStoredType();
+    Preconditions.checkArgument(
+        dataType == FieldSpec.DataType.LONG || dataType == FieldSpec.DataType.INT,
+        "MAXLONG aggregation function can only be applied to columns of integer types");
+    Dictionary dictionary = dataSource.getDictionary();
+    if (dictionary != null) {
+      return ((Number) dictionary.getMaxVal()).longValue();
+    }
+    return ((Number) dataSource.getDataSourceMetadata().getMaxValue()).longValue();
+  }
+
+  private static Double toDouble(Comparable<?> value) {
+    if (value instanceof Double) {
+      return (Double) value;
+    } else if (value instanceof Number) {
+      return ((Number) value).doubleValue();
+    } else {
+      return Double.parseDouble(value.toString());
+    }
+  }
+
+  private static Set getDistinctValueSet(Dictionary dictionary, String explainName) {
+    int dictionarySize = dictionary.length();
+    switch (dictionary.getValueType()) {
+      case INT:
+        IntOpenHashSet intSet = new IntOpenHashSet(dictionarySize);
+        for (int dictId = 0; dictId < dictionarySize; dictId++) {
+          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, explainName);
+          intSet.add(dictionary.getIntValue(dictId));
+        }
+        return intSet;
+      case LONG:
+        LongOpenHashSet longSet = new LongOpenHashSet(dictionarySize);
+        for (int dictId = 0; dictId < dictionarySize; dictId++) {
+          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, explainName);
+          longSet.add(dictionary.getLongValue(dictId));
+        }
+        return longSet;
+      case FLOAT:
+        FloatOpenHashSet floatSet = new FloatOpenHashSet(dictionarySize);
+        for (int dictId = 0; dictId < dictionarySize; dictId++) {
+          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, explainName);
+          floatSet.add(dictionary.getFloatValue(dictId));
+        }
+        return floatSet;
+      case DOUBLE:
+        DoubleOpenHashSet doubleSet = new DoubleOpenHashSet(dictionarySize);
+        for (int dictId = 0; dictId < dictionarySize; dictId++) {
+          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, explainName);
+          doubleSet.add(dictionary.getDoubleValue(dictId));
+        }
+        return doubleSet;
+      case BIG_DECIMAL:
+        ObjectOpenHashSet<BigDecimal> bigDecimalSet = new ObjectOpenHashSet<>(dictionarySize);
+        for (int dictId = 0; dictId < dictionarySize; dictId++) {
+          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, explainName);
+          bigDecimalSet.add(dictionary.getBigDecimalValue(dictId));
+        }
+        return bigDecimalSet;
+      case STRING:
+        ObjectOpenHashSet<String> stringSet = new ObjectOpenHashSet<>(dictionarySize);
+        for (int dictId = 0; dictId < dictionarySize; dictId++) {
+          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, explainName);
+          stringSet.add(dictionary.getStringValue(dictId));
+        }
+        return stringSet;
+      case BYTES:
+        ObjectOpenHashSet<ByteArray> bytesSet = new ObjectOpenHashSet<>(dictionarySize);
+        for (int dictId = 0; dictId < dictionarySize; dictId++) {
+          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(dictId, explainName);
+          bytesSet.add(new ByteArray(dictionary.getBytesValue(dictId)));
+        }
+        return bytesSet;
+      default:
+        throw new IllegalStateException();
+    }
+  }
+
+  private static HyperLogLog getDistinctValueHLL(Dictionary dictionary, int log2m, String explainName) {
+    HyperLogLog hll = new HyperLogLog(log2m);
+    int length = dictionary.length();
+    for (int i = 0; i < length; i++) {
+      QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, explainName);
+      hll.offer(dictionary.get(i));
+    }
+    return hll;
+  }
+
+  private static UltraLogLog getDistinctValueULL(Dictionary dictionary, int p, String explainName) {
+    UltraLogLog ull = UltraLogLog.create(p);
+    int length = dictionary.length();
+    for (int i = 0; i < length; i++) {
+      QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, explainName);
+      Object value = dictionary.get(i);
+      UltraLogLogUtils.hashObject(value).ifPresent(ull::add);
+    }
+    return ull;
+  }
+
+  private static HyperLogLogPlus getDistinctValueHLLPlus(Dictionary dictionary, int p, int sp, String explainName) {
+    HyperLogLogPlus hllPlus = new HyperLogLogPlus(p, sp);
+    int length = dictionary.length();
+    for (int i = 0; i < length; i++) {
+      QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, explainName);
+      hllPlus.offer(dictionary.get(i));
+    }
+    return hllPlus;
+  }
+
+  private static HyperLogLog getDistinctCountHLLResult(Dictionary dictionary,
+      DistinctCountHLLAggregationFunction function, String explainName) {
+    if (dictionary.getValueType() == FieldSpec.DataType.BYTES) {
+      // Treat BYTES value as serialized HyperLogLog
+      try {
+        QueryThreadContext.checkTerminationAndSampleUsage(explainName);
+        HyperLogLog hll = ObjectSerDeUtils.HYPER_LOG_LOG_SER_DE.deserialize(dictionary.getBytesValue(0));
+        int length = dictionary.length();
+        for (int i = 1; i < length; i++) {
+          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, explainName);
+          hll.addAll(ObjectSerDeUtils.HYPER_LOG_LOG_SER_DE.deserialize(dictionary.getBytesValue(i)));
+        }
+        return hll;
+      } catch (Exception e) {
+        throw new RuntimeException("Caught exception while merging HyperLogLogs", e);
+      }
+    } else {
+      return getDistinctValueHLL(dictionary, function.getLog2m(), explainName);
+    }
+  }
+
+  private static HyperLogLogPlus getDistinctCountHLLPlusResult(Dictionary dictionary,
+      DistinctCountHLLPlusAggregationFunction function, String explainName) {
+    if (dictionary.getValueType() == FieldSpec.DataType.BYTES) {
+      // Treat BYTES value as serialized HyperLogLogPlus
+      try {
+        QueryThreadContext.checkTerminationAndSampleUsage(explainName);
+        HyperLogLogPlus hllplus = ObjectSerDeUtils.HYPER_LOG_LOG_PLUS_SER_DE.deserialize(dictionary.getBytesValue(0));
+        int length = dictionary.length();
+        for (int i = 1; i < length; i++) {
+          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, explainName);
+          hllplus.addAll(ObjectSerDeUtils.HYPER_LOG_LOG_PLUS_SER_DE.deserialize(dictionary.getBytesValue(i)));
+        }
+        return hllplus;
+      } catch (Exception e) {
+        throw new RuntimeException("Caught exception while merging HyperLogLogPluses", e);
+      }
+    } else {
+      return getDistinctValueHLLPlus(dictionary, function.getP(), function.getSp(), explainName);
+    }
+  }
+
+  private static Object getDistinctCountSmartHLLResult(Dictionary dictionary,
+      DistinctCountSmartHLLAggregationFunction function, String explainPlanName) {
+    if (dictionary.length() > function.getThreshold()) {
+      // Store values into a HLL when the dictionary size exceeds the conversion threshold
+      return getDistinctValueHLL(dictionary, function.getLog2m(), explainPlanName);
+    } else {
+      return getDistinctValueSet(dictionary, explainPlanName);
+    }
+  }
+
+  private static Object getDistinctCountSmartHLLPlusResult(Dictionary dictionary,
+      DistinctCountSmartHLLPlusAggregationFunction function, String explainName) {
+    if (dictionary.length() > function.getThreshold()) {
+      // Store values into a HLLPlus when the dictionary size exceeds the conversion threshold
+      return getDistinctValueHLLPlus(dictionary, function.getP(), function.getSp(), explainName);
+    } else {
+      return getDistinctValueSet(dictionary, explainName);
+    }
+  }
+
+  private static UltraLogLog getDistinctCountULLResult(Dictionary dictionary,
+      DistinctCountULLAggregationFunction function, String explainName) {
+    if (dictionary.getValueType() == FieldSpec.DataType.BYTES) {
+      // Treat BYTES value as serialized UltraLogLog and merge
+      try {
+        QueryThreadContext.checkTerminationAndSampleUsage(explainName);
+        UltraLogLog ull = ObjectSerDeUtils.ULTRA_LOG_LOG_OBJECT_SER_DE.deserialize(dictionary.getBytesValue(0));
+        int length = dictionary.length();
+        for (int i = 1; i < length; i++) {
+          QueryThreadContext.checkTerminationAndSampleUsagePeriodically(i, explainName);
+          ull.add(ObjectSerDeUtils.ULTRA_LOG_LOG_OBJECT_SER_DE.deserialize(dictionary.getBytesValue(i)));
+        }
+        return ull;
+      } catch (Exception e) {
+        throw new RuntimeException("Caught exception while merging UltraLogLogs", e);
+      }
+    } else {
+      return getDistinctValueULL(dictionary, function.getP(), explainName);
+    }
+  }
+
+  private static Object getDistinctCountSmartULLResult(Dictionary dictionary,
+      DistinctCountSmartULLAggregationFunction function, String explainName) {
+    if (dictionary.length() > function.getThreshold()) {
+      return getDistinctValueULL(dictionary, function.getP(), explainName);
+    } else {
+      return getDistinctValueSet(dictionary, explainName);
+    }
   }
 }
